+++ test/lib/index.spec.js
@@ -230,76 +230,6 @@ describe('The API', function() {
 
 describe('Mocking the file system', function() {
 
-  describe('fs.lstat(path, callback)', function() {
-    beforeEach(function() {
-      mock({
-        'file.txt': mock.file({
-          content: 'content',
-          mtime: new Date(1)
-        }),
-        link: mock.symlink({
-          path: './file.txt',
-          mtime: new Date(2)
-        })
-      });
-    });
-    afterEach(mock.restore);
-
-    it('stats a symbolic link', function(done) {
-      fs.lstat('link', function(err, stats) {
-        if (err) {
-          return done(err);
-        }
-        assert.isTrue(stats.isSymbolicLink());
-        assert.isFalse(stats.isFile());
-        assert.equal(stats.mtime.getTime(), 2);
-        done();
-      });
-    });
-
-    it('stats a regular file', function(done) {
-      fs.lstat('file.txt', function(err, stats) {
-        if (err) {
-          return done(err);
-        }
-        assert.isTrue(stats.isFile());
-        assert.isFalse(stats.isSymbolicLink());
-        assert.equal(stats.mtime.getTime(), 1);
-        done();
-      });
-    });
-  });
-
-  describe('fs.lstatSync(path)', function() {
-    beforeEach(function() {
-      mock({
-        'file.txt': mock.file({
-          content: 'content',
-          mtime: new Date(1)
-        }),
-        link: mock.symlink({
-          path: './file.txt',
-          mtime: new Date(2)
-        })
-      });
-    });
-    afterEach(mock.restore);
-
-    it('stats a symbolic link', function() {
-      const stats = fs.lstatSync('link');
-      assert.isTrue(stats.isSymbolicLink());
-      assert.isFalse(stats.isFile());
-      assert.equal(stats.mtime.getTime(), 2);
-    });
-
-    it('stats a regular file', function() {
-      const stats = fs.lstatSync('file.txt');
-      assert.isTrue(stats.isFile());
-      assert.isFalse(stats.isSymbolicLink());
-      assert.equal(stats.mtime.getTime(), 1);
-    });
-  });
-
   describe('fs.realpath(path, [cache], callback)', function() {
     // based on binding.lstat and binding.readlink so tested elsewhere as well
